requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
beautifulsoup4
blinker
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
cssselect2
Cython
dataclasses-json
decorator
distro
dtw-python
edge-tts
exceptiongroup
filelock
fonttools
frozenlist
fsspec
gitdb
GitPython
greenlet
h11
html5lib
httpcore
httpx
idna
imageio
imageio-ffmpeg
importlib_metadata
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchainhub
langgraph
langsmith
llvmlite
Markdown
markdown-it-py
MarkupSafe
marshmallow
mdurl
more-itertools
moviepy
mpmath
multidict
mypy-extensions
narwhals
networkx
numba
numexpr
numpy
openai
openai-whisper
orjson
packaging
pandas
pdfkit
pillow
pip
proglog
protobuf
pyarrow
pycparser
pydantic
pydantic_core
pydeck
pydyf
Pygments
pyphen
PySocks
python-dateutil
python-decouple
pytz
PyYAML
referencing
regex
reportlab
requests
rich
rpds-py
scipy
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
streamlit
sympy
tenacity
tiktoken
tinycss2
toml
torch
tornado
tqdm
types-requests
typing_extensions
typing-inspect
tzdata
urllib3
watchdog
weasyprint
webencodings
wheel
whisper-timestamped
win-inet-pton
yarl
zipp
zopfli
